[ios] Pin JS and URL loading to the NetworkPolicy
So when the wifi disabled the PP HTML with https content will not be loaded if the user selects Never/Always ask (prompt will be shown)
